Retarget to net6.0 because of the long-term support; version bump
