+++ Xamarin/XamCtlTest/XamCtlTest.iOS/XamCtlTest.iOS.csproj
@@ -14,10 +14,10 @@
     <MtouchEnableSGenConc>true</MtouchEnableSGenConc>
     <MtouchHttpClientHandler>NSUrlSessionHandler</MtouchHttpClientHandler>
     <ProvisioningType>automatic</ProvisioningType>
-    <AssemblyVersion>8.0.0.1038</AssemblyVersion>
-    <FileVersion>8.0.0.1038</FileVersion>
-    <PreviousVersion>8.0.0.1037</PreviousVersion>
-    <Version>8.0.0.1038</Version>
+    <AssemblyVersion>8.0.0.1040</AssemblyVersion>
+    <FileVersion>8.0.0.1040</FileVersion>
+    <PreviousVersion>8.0.0.1039</PreviousVersion>
+    <Version>8.0.0.1040</Version>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|iPhoneSimulator' ">
     <DebugSymbols>true</DebugSymbols>
